requirements.txt
flask
google-generativeai
Pillow
werkzeug
python-dotenv
